pyproject.toml
[tool.poetry]
name = ""
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"

pip = "^25.0.1"
ipython = "^8.32.0"

black = { version = "^25.1.0", extras = ["d"] }
isort = "^6.0.0"
flake8 = "^7.1.2"
autoflake = "^2.3.1"

matplotlib = "^3.10.0"

setuptools = "^75.8.0"
pytest-mock = "^3.14.0"
pybind11 = "^2.13.6"
poetry-core = "^2.1.2"

pint = "^0.24.4"
pandas = "^2.2.3"
colorama = "^0.4.6"
beartype = "^0.20.0"

jax = { version = "==0.5.0", extras = ["cuda"] }
oryx = "^0.2.9"
equinox = "^0.11.11"
sympy2jax = "^0.0.6"
jaxtyping = "^0.2.38"
optax = "^0.2.4"

# pintax = { git = "https://github.com/Alan-Chen99/pintax.git" }
jsonpickle = "^4.0.5"

[tool.isort]
profile = "black"
known_first_party = ["pintax", "lib"]

[tool.pyright]
exclude = ["**/result*", ".cache"]
typeCheckingMode = "strict"

reportAny = false
reportExplicitAny = false
reportIgnoreCommentWithoutRule = false
reportImplicitOverride = false
reportImportCycles = false
reportMissingParameterType = false
reportMissingTypeArgument = false
reportMissingTypeStubs = false
reportPrivateImportUsage = false
reportPrivateUsage = false
reportUnannotatedClassAttribute = false
reportUnknownArgumentType = false
reportUnknownLambdaType = false
reportUnknownMemberType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportUnnecessaryIsInstance = false
reportUnusedImport = false
reportUnusedParameter = false

reportUnusedFunction = "warning"
reportUnusedVariable = "warning"
